aws-sdk/pyproject.toml
[tool.poetry]
name = "aws-sdk"
version = "0.1.0"
description = ""
authors = ["TonyChou <cool21540125@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = {extras = ["crt"], version = "^1.34.14"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
